+++ techpourtoutes/forms/__init__.py
@@ -1,13 +1,11 @@
 from .account_edit_form import AccountEditForm
+from .engagement_forms import EngagementForm, TrainingAmbassadorForm, WorkshopForm
 from .login_request_form import LoginRequestForm
-from .pro_form import ProForm
-from .training_ambassador_form import TrainingAmbassadorForm
-from .workshop_form import WorkshopForm
 
 __all__ = [
     "AccountEditForm",
+    "EngagementForm",
     "LoginRequestForm",
-    "ProForm",
     "TrainingAmbassadorForm",
     "WorkshopForm",
 ]
