+++ techpourtoutes/forms/engagement_forms/__init__.py
@@ -1,0 +1,9 @@
+from .engagement_form import EngagementForm
+from .training_ambassador_form import TrainingAmbassadorForm
+from .workshop_form import WorkshopForm
+
+__all__ = [
+    "EngagementForm",
+    "TrainingAmbassadorForm",
+    "WorkshopForm",
+]
