+++ techpourtoutes/tests/test_forms.py
@@ -3,40 +3,40 @@ import pytest
 
 @pytest.mark.django_db
 def test_pro_form_valid(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data=valid_pro_data)
+    form = EngagementForm(data=valid_pro_data)
     assert form.is_valid(), form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_missing_required_field(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data={**valid_pro_data, "phone": ""})
+    form = EngagementForm(data={**valid_pro_data, "phone": ""})
     assert not form.is_valid()
     assert "phone" in form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_duplicate_email(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data=valid_pro_data)
+    form = EngagementForm(data=valid_pro_data)
     assert form.is_valid()
     form.save()
 
-    form2 = ProForm(data=valid_pro_data)
+    form2 = EngagementForm(data=valid_pro_data)
     assert not form2.is_valid()
     assert "email" in form2.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_save_creates_pro(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
     from techpourtoutes.models import Pro
 
-    form = ProForm(data=valid_pro_data)
+    form = EngagementForm(data=valid_pro_data)
     assert form.is_valid()
     pro = form.save()
     assert isinstance(pro, Pro)
@@ -45,75 +45,75 @@ def test_pro_form_save_creates_pro(valid_pro_data):
 
 @pytest.mark.django_db
 def test_pro_form_invalid_professional_situation(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data={**valid_pro_data, "professional_situation": "not-a-choice"})
+    form = EngagementForm(data={**valid_pro_data, "professional_situation": "not-a-choice"})
     assert not form.is_valid()
     assert "professional_situation" in form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_blank_professional_situation(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data={**valid_pro_data, "professional_situation": ""})
+    form = EngagementForm(data={**valid_pro_data, "professional_situation": ""})
     assert not form.is_valid()
     assert "professional_situation" in form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_requires_terms_accepted(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data={**valid_pro_data, "terms_accepted": False})
+    form = EngagementForm(data={**valid_pro_data, "terms_accepted": False})
     assert not form.is_valid()
     assert "terms_accepted" in form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_invalid_civility(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data={**valid_pro_data, "civility": "Mademoiselle"})
+    form = EngagementForm(data={**valid_pro_data, "civility": "Mademoiselle"})
     assert not form.is_valid()
     assert "civility" in form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_structure_name_required_when_working(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(data={**valid_pro_data, "structure_name": ""})
+    form = EngagementForm(data={**valid_pro_data, "structure_name": ""})
     assert not form.is_valid()
     assert "structure_name" in form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_structure_name_not_required_when_retired(valid_pro_data):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
     data = {
         **valid_pro_data,
         "professional_situation": "retired",
         "structure_name": "",
     }
-    form = ProForm(data=data)
+    form = EngagementForm(data=data)
     assert form.is_valid(), form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_with_pro_email_is_disabled(pro):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(pro=pro)
+    form = EngagementForm(pro=pro)
     assert form.fields["email"].disabled
 
 
 @pytest.mark.django_db
 def test_pro_form_with_pro_sets_initial_values(pro):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
-    form = ProForm(pro=pro)
+    form = EngagementForm(pro=pro)
     assert form.initial["email"] == pro.email
     assert form.initial["first_name"] == pro.first_name
     assert form.initial["last_name"] == pro.last_name
@@ -121,7 +121,7 @@ def test_pro_form_with_pro_sets_initial_values(pro):
 
 @pytest.mark.django_db
 def test_pro_form_with_pro_terms_not_required(pro):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
     data = {
         "civility": pro.civility,
@@ -134,13 +134,13 @@ def test_pro_form_with_pro_terms_not_required(pro):
         "job_title": pro.job_title,
         "structure_name": pro.structure_name,
     }
-    form = ProForm(data=data, pro=pro)
+    form = EngagementForm(data=data, pro=pro)
     assert form.is_valid(), form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_with_pro_allows_own_email(pro):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
 
     data = {
         "civility": pro.civility,
@@ -154,13 +154,13 @@ def test_pro_form_with_pro_allows_own_email(pro):
         "structure_name": pro.structure_name,
         "terms_accepted": True,
     }
-    form = ProForm(data=data, pro=pro)
+    form = EngagementForm(data=data, pro=pro)
     assert form.is_valid(), form.errors
 
 
 @pytest.mark.django_db
 def test_pro_form_with_pro_save_updates_in_place(pro):
-    from techpourtoutes.forms import ProForm
+    from techpourtoutes.forms import EngagementForm
     from techpourtoutes.models import Pro
 
     data = {
@@ -175,7 +175,7 @@ def test_pro_form_with_pro_save_updates_in_place(pro):
         "structure_name": pro.structure_name,
         "terms_accepted": True,
     }
-    form = ProForm(data=data, pro=pro)
+    form = EngagementForm(data=data, pro=pro)
     assert form.is_valid(), form.errors
     saved = form.save()
